UserLoginUpdate
Connor G and Mike W worked on this. This takes an Email and Password from homepage.html, checks if its a valid account in Users. Then checks what type of user it is to be redirected to the correct login page(Admin, parent or student). I also updated the DB2Pop.sql to have the correct admin id for the default admin account.