Duplicate Meeting Signup +DB2Pop.sql update
Mike W Updated the DB2Pop.sql. Created radio buttons so when joining a meeting as mentor or mentee  you join the single meeting as indicated by the ID you entered. Or if reoccuring is selected then they will be enrolled in  all meetings with the same name (Ie reoccuring meetings).